JSon : Warning fix
git-svn-id: http://svn.ta3d.org/yuni/trunk@2363 7e728022-c0ab-4603-8e65-c666ead28a33

+++ src/yuni/private/marshal/json/grammar.cpp
@@ -937,7 +937,7 @@ namespace JSON
 			Private::JSON::grammar.symbols.table[symbol].kind == Private::JSON::Symbols::kdNonTerminal;
 
 		childIndex = 0;
-		childrenCount = token.tokens.size();
+		childrenCount = (uint) token.tokens.size();
 	}
